refactor: remove dead code and unused resources
- Remove unnecessary i32 cast in Rust JNI layer
- Delete unused string resources from strings.xml
- Verify JNI integrity and project structure

+++ rust/src/lib.rs
@@ -122,7 +122,7 @@ jni_fn! {
         let array = env.new_object_array(cores, &class, initial_element)?;
 
         for i in 1..cores {
-            let governor = kernel::cpu::get_core_governor(i as i32);
+            let governor = kernel::cpu::get_core_governor(i);
             let s = env.new_string(governor)?;
             array.set_element(env, i as usize, s)?;
         }
